Refactor : 장소 업데이트 부분 Refactoring
장소 업데이트 부분에 MVVM 모델을 적용하였고,
사진이 추가될 때, 위치와 시간이 변하는 것을 개선하였다.
